requirements.txt
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.8
flake8==5.0.4
flake8-builtins==2.5.0
flake8-commas==4.0.0
flake8-docstrings==1.6.0
flake8-isort==6.1.2
flake8-multiline-containers==0.0.19
flake8-polyfill==1.0.2
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pep8-naming==0.14.1
platformdirs==4.3.6
pluggy==1.5.0
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pyTelegramBotAPI==4.14.1
pytest==7.1.3
pytest-timeout==2.1.0
python-dotenv==0.20.0
requests==2.26.0
snowballstemmer==2.2.0
tomli==2.0.2
typing_extensions==4.12.2
urllib3==1.26.20
